Refactor executions exporter to own file
